Fix interpolation spline issues
It seems that one can't reset the state of the vtkKochanekSpline
once you have used it, so instead we recreate new instances
for each interpolation of the slice control points.  Not a huge
performance difference - if performance becomes and issue we can
cache the interpolation curves rather than recreating them
for each slice optimization.