ENH: new threshold in segmented operation
This allows the threshold to be applied within previously
segmented regions.  For example, if user has defined several
regions in one label map, a new label map can be made
where the threshold is applied with in each region.  An
application might be to use erode and save island to
disconnect a region, and then dilate to include the full
extent; then a threshold in segmented would restrore the
nice boundaries without connection through the leak.