Add hermite interoplation option using edge tangents
Does a great job where the edge is well defined,
and fewer control points can give a much better
curve definition.  But when the control point isn't
on a good edge the edge tangent is noisy and hard
to control.  May need to add a user mode to turn off
edge tangent calculation on a per-control point level