ENH: new 'auto view' mode and 'return to last jump' mode
Now whenever you use the 'j' or 'k' keys to jump to the nearest control
point for a slice, the slice you *jumped from* is remembered.  You
can use the 'i' key to go back to the place you jumped from.  This is
intended to simplify the process of checking nearby curves before selecting
the new slice to define.

As a helper for this, the 'I' key now does a little animation where it
jumps for 100ms to the previous slice, then back to the current, then
to the next slice and then back to the current.  This is meant to provide
a quick visual doublecheck of the consistency of curves.